Proyecto "completado"

+++ src/models/CompararInt.java
@@ -6,7 +6,7 @@ public class CompararInt implements Comparator<Integer> {
 
 	@Override
 	public int compare(Integer o1, Integer o2) {
-		return (int) (o1 - o2);
+		return o1 - o2;
 	}
 
 }
